pyproject.toml
[project]
name = "evo-algo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "gradio>=4.0.0",
    "ipython>=9.8.0",
    "flask>=3.1.2",
    "numba>=0.63.1",
]
